+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+- Add mixins for [\[UNOFFICIAL\] TaCZ NeoForge Port](https://www.curseforge.com/minecraft/mc-mods/tacz-1-21-1) 1.1.8+ [#22](https://github.com/gizmo-ds/taczjs-mod/issues/22)
+
 ### v1.4.0+mc1.21.1 - 2026-01-11
 
 - Added [\[UNOFFICIAL\] TaCZ NeoForge Port](https://www.curseforge.com/minecraft/mc-mods/tacz-1-21-1) support